Directions

Correlation functions can be computed in many directions depending on how test line segments are aligned with the input array. The default directions are [:x] for 1D, [:x, :y] for 2D and [:x, :y, :z] for 3D arrays. Possible directions and their meaning are described in the documentation for directionNDp functions where N stands for 1, 2 and 3.

CorrelationFunctions.direction1Dp — Function

direction1Dp(sym)

Return true if sym is 1D direction or false otherwise.

The only possible 1D direction is :x.

CorrelationFunctions.direction2Dp — Function

direction2Dp(sym)

Return true if sym is 2D direction or false otherwise.

Known directions are:

:x and :y. Correlation functions are computed along unit vectors (1, 0) and (0, 1) respectively.

:xy_main and :xy_anti. Correlation functions are computed in diagonal directions (1, 1) and (-1, 1) respectively.

CorrelationFunctions.direction3Dp — Function

direction3Dp(sym)

Return true if sym is 3D direction or false otherwise.

Known directions are:

:x, :y and :z. Correlation functions are computed along unit vectors (1, 0, 0), (0, 1, 0) and (0, 0, 1) respectively.

:xy_main, :xz_main and :yz_main. Correlation functions are computed in diagonal directions (1, 1, 0), (1, 0, 1) and (0, 1, 1) respectively.

:xy_anti, :xz_anti and :yz_anti. Correlation functions are computed in diagonal directions (-1, 1, 0), (-1, 0, 1) and (0, -1, 1) respectively.

From :diag1 to :diag4. Corresponding directions are (1, 1, 1), (-1, 1, 1), (1, -1, 1) and (1, 1, -1).