Functions

The following correlation functions are supported:

Lineal-path $L_2$ function.

Two point $S_2$ function.

Cluster $C_2$ function.

Surface-surface $F_{ss}$ function.

Surface-void $F_{sv}$ function.

Pore size $P$ function.

Chord length $p$ function.

CorrelationFunctions.l2 — Function

l2(array, phase; [len][, directions,] periodic = false)

Calculate L₂ (lineal path) correlation function for one-, two- or three-dimensional multiphase system.

L₂(x) equals to probability that all elements of a line segment with length x cut from the array belong to the same phase. This implementation calculates L₂(x) for all xes in the range from 1 to len which defaults to half of the minimal dimension of the array.

Examples

julia> l2([1,1,1,0,1,1], 1; len = 6)[:x]
6-element Array{Float64,1}:
 0.8333333333333334
 0.6
 0.25
 0.0
 0.0
 0.0

For a list of possible dimensions, see also: direction1Dp, direction2Dp, direction3Dp.

CorrelationFunctions.s2 — Function

s2(array, phase[; len][, directions,] periodic = false)
s2(array, SeparableIndicator(χ₁, χ₂)[; len][,directions,] periodic = false)
s2(array, InseparableIndicator(χ)[; len][,directions,] periodic = false)

Calculate S₂ (two point) correlation function for one-, two- or three-dimensional multiphase system.

S₂(x) equals to probability that corner elements of a line segment with the length x cut from the array belong to the same phase. This implementation calculates S₂(x) for all xes in the range from 1 to len which defaults to half of the minimal dimenstion of the array.

More generally, you can provide indicator function χ instead of phase. In this case S₂ function calculates probability of χ(x, y) returing true where x and y are two corners of a line segment. Indicator functions must be wrapped in either SeparableIndicator or InseparableIndicator. Some computations for separable indicator functions are optimized.

Examples

julia> s2([1,1,1,0,1,1], 1; len = 6)[:x]
6-element Array{Float64,1}:
 0.8333333333333334
 0.6
 0.5
 0.6666666666666666
 1.0
 1.0

See also: direction1Dp, direction2Dp, direction3Dp, SeparableIndicator, InseparableIndicator.

CorrelationFunctions.c2 — Function

c2(array, phase[; len,][directions,] periodic = false)

Calculate C₂ (cluster) correlation function for one-, two- or three-dimensional multiphase system.

C₂(x) equals to probability that corner elements of a line segment with the length x cut from the array belong to the same cluster of the specific phase. This implementation calculates C2 for all xes in the range from 1 to len which defaults to half of the minimal dimension of the array.

Examples

julia> c2([1,1,1,0,1,1], 1; len = 6)[:x]
6-element Array{Float64,1}:
 0.8333333333333334
 0.6
 0.25
 0.0
 0.0
 0.0

For a list of possible dimensions, see also: direction1Dp, direction2Dp, direction3Dp.

CorrelationFunctions.surfsurf — Function

surfsurf(array, phase[; len,][directions,] periodic = false)

Calculate Fss(x) (surface-surface) correlation function for one-, two- or three-dimensional multiphase system.

Fss(x) equals to probability that corner elements of a line segment with the length x cut from the array belong to the boundary of a cluster with the phase phase. This implementation calculates surface-surface function for all xs in the range from 1 to len which defaults to half of the minimal dimension of the array.

For a list of possible dimensions, see also: direction1Dp, direction2Dp, direction3Dp.

CorrelationFunctions.surfvoid — Function

surfvoid(array, phase[; len,][directions,] periodic = false)

Calculate Fsv(x) (surface-void) correlation function for one-, two- or three-dimensional multiphase system.

Fsv(x) equals to probability that one corner of a line segment with the length x cut from the array belongs to the boundary of a cluster with the phase phase and the other belongs to the void phase 0. This implementation calculates surface-void function for all xs in the range from 1 to len which defaults to half of the minimal dimension of the array.

For a list of possible dimensions, see also: direction1Dp, direction2Dp, direction3Dp.

CorrelationFunctions.pore_size — Function

pore_size(array, phase = 0; nbins = 10)

Calculate pore size correlation function for one-, two- or three-dimensional multiphase system.

Pore size correlation function P(x) equals to probability of inserting a ball with radius R ∈ [x, x + δx] into a system so that it lies entirely in the phase phase.

This implementation divides the range of possible radii into nbins subranges and returns a normalized histogram of radii. This is roughly equal to integrating P(x) for each subrange.

CorrelationFunctions.chord_length — Function

chord_length(array, phase[; directions,] nbins = 10)

Calculate chord length correlation function for one-, two- or three-dimensional multiphase system.

Cord length function p(x) equals to probability of finding a chord whose length is in the range [x, x+δx] and which lies entirely in the phase phase. A chord is a line segment which touches the boundary of a same-phase cluster with its ends.

This implementation bins chord lengths into nbins bins and returns normalized histogram on collected data and the mean chord length in a tuple.

Examples

julia> chord_length([1, 0, 0, 0, 0, 1], 0)
(StatsBase.Histogram{Float64,1,Tuple{StepRangeLen{Float64,Base.TwicePrecision{Float64},Base.TwicePrecision{Float64}}}}
edges:
  4.0:1.0:5.0
weights: [1.0]
closed: left
isdensity: false, 4.0)

For a list of possible dimensions, see also: direction1Dp, direction2Dp, direction3Dp.