Indicator Functions

Internally, the functions c2, surfsurf and surfvoid (see Functions) are reduced to s2 passing more generic indicator functions rather than simply a phase. This feature is also exposed to users. If you want to use a custom indicator function, you need to wrap it to either SeparableIndicator or InseparableIndicator structure, calling the corresponding constructor. Note that s2 performs much better on big arrays when using SeparableIndicator.

CorrelationFunctions.AbstractIndicator — Type

Abstract type for indicator functions $\mathbb{R}^{2n} \rightarrow \left\{0, 1\right\}$ where $n = 1, 2 \text{ or } 3$.

CorrelationFunctions.SeparableIndicator — Type

SeparableIndicator(χ₁, χ₂)

Type for separable indicator function, that is for such an indicator function which can be written as $\chi(x,y) = \chi_1(x)\chi_2(y)$.

χ1 and χ2 must be functions of one argument which return a value of Bool type.

NB: This indicator function is not symmetric (i.e. $\chi(x,y) \ne \chi(y,x)$). This behaviour is intentional. For example you can write such an indicator, so the corresponding correlation function is sensitive to the spatial orientation of a system.

"That one, too fat! This one, too tall! This one… too symmetrical!"

CorrelationFunctions.InseparableIndicator — Type

InseparableIndicator(χ)

Type for inseparable indicator function, that is for such an indicator function which cannot be written as $\chi(x,y) = \max(\Chi(x, y), \Chi(y, x))$, where $\Chi(x, y) = \chi_1(x)\chi_2(y)$.

χ must be a function of two arguments which returns a value of Bool type.