Correlation Maps

CorrelationFunctions.Map.l2 — Function

l2(image; periodic = false)

Calculate L₂ (lineal path) correlation function map for the 1,2,3-dimensional image and return a CFMap object.

The image contains the probability of the voxel being in the correct phase.

For L2, only probabilities 0 and 1 are currently implemented.

Examples

julia> l2([1 0; 0 1]; periodic=true).result
3×2 Matrix{Float64}:
 0.0  0.5
 0.5  0.0
 0.0  0.5

CorrelationFunctions.Map.s2 — Function

s2(image; periodic = false)

Calculate S₂ (two point) correlation function map for the N-dimensional image and return a CFMap object.

The image contains the probability of the voxel being in the correct phase.

Examples

julia> s2([1 0; 0 1]; periodic=true).result
2×2 Matrix{Float64}:
 0.5  0.0
 0.0  0.5

CorrelationFunctions.Map.c2 — Function

c2(image; periodic = false)

Calculate C₂ (cluster) correlation function map for the N-dimensional image and return a CFMap object.

The image contains the probability of the voxel being in the correct phase.

Examples

julia> c2([1 0; 0 1]; periodic=true).result
2×2 Matrix{Float64}:
 0.5  0.0
 0.0  0.0

CorrelationFunctions.Map.surfsurf — Function

surfsurf(image; periodic = false)

Calculate Fss(x) (surface-surface) correlation function map for the N-dimensional image and return a CFMap object.

The image contains the probability of the voxel being in the correct phase.

Examples

julia> surfsurf([1 0; 0 1]; periodic=true).result
2×2 Matrix{Float64}:
 0.125  0.125
 0.125  0.125

CorrelationFunctions.Map.surfvoid — Function

surfvoid(image; periodic = false)

Calculate Fsv(x) (surface-void) correlation function map for the N-dimensional image and return a CFMap object.

The image contains the probability of the voxel being in the correct phase.

Examples

julia> surfvoid([1 0; 0 1]; periodic=true).result
3×3 Matrix{Float64}:
 0.176777  0.176777  0.176777
 0.176777  0.176777  0.176777
 0.176777  0.176777  0.176777