Utilities

CorrelationFunctions.Utilities.read_cuboid — Function

read_cuboid(cfgpath :: String)

Read 3D array from a disk. The data on disk consists of two files: JSON configuration file (which is passed to this function) and a raw binary array data.

Scheme of the configuration file is as follows, where x, y and z are dimensions of the array:

{
    "dimensions": [x, y, z],
    "datapath": "file-with-data"
}

The file with binary data, whose name is specified in datapath field, is searched relatively to the directory with the JSON configuration file. Its size must be exactly x⋅y⋅z bytes, each byte containing an element of the resulting array.

read_cuboid(datapath :: String, side, dim)

Read 3D array from a disk. The data on the disk must be in binary format, one octet per sample. Totally, there must be $side^{dim}$ octets which are read into $side \times side \times \dots \times side$ array.

CorrelationFunctions.Utilities.lowfreq_energy_ratio — Function

lowfreq_energy_ratio(array, fraction = 0.5)

Calculate a ratio $E_a/E$ where $E$ is a total energy of a signal array and $E_a$ is the energy concentrated in frequencies $[0, af/2]$ where $f$ is the sampling rate and $a$ is set via parameter fraction. mean(array) is subtracted from the array before calculations.

This function can be helpful in estimating if array is suitable for calculating surface-surface or surface-void function. An empirical criterion is that if this function returns a value greater than 0.97, the array is good.

CorrelationFunctions.Utilities.extract_edges — Function

extract_edges(array, filter)

Perform edge extraction in the same way as in surfsurf and surfvoid functions from Map and Directional modules. array may be a CUDA array or an ordinary array. filter is a value of EdgeFilter type which selects an edge extraction algorithm.

See also: EdgeFilter, BoundaryConditions.

CorrelationFunctions.Utilities.BoundaryConditions — Type

BoundaryConditions

Abstract type which specifies boundary conditions for edge extraction filter.

See also: BCPeriodic, BCReflect.

CorrelationFunctions.Utilities.BCPeriodic — Type

BCPeriodic()

The signal is extended over borders periodically. These conditions have implementation for CPU and GPU and are default if periodic argument to surface functions is true.

See also: BoundaryConditions, extract_edges.

CorrelationFunctions.Utilities.BCReflect — Type

BCReflect()

The signal is extended over borders by reflecting relative to the edge. This is a default setting when periodic argument to surface functions is false.

NB: It's natural to assume zero padding must be used for non-periodic mode. But we want an equality $F_{ss}^{(false)} = F_{ss}^{(true)}$ for two-phase system, hence we cannot pad the input with some constant value.

See also: BoundaryConditions, extract_edges.

CorrelationFunctions.Utilities.FilterKernel — Type

FilterKernel

Abstract type for one of edge detecting filters.

See also: Kernel3x3, Kernel5x5.

CorrelationFunctions.Utilities.Kernel3x3 — Type

Kernel3x3()

Make a suboptimal 3x3 edge detection filter. This filter must be used only for images with insufficient resolution (lowfreq_energy_ratio(array) ≈ 0.97).

See also: FilterKernel, extract_edges.

CorrelationFunctions.Utilities.Kernel5x5 — Type

Kernel5x5()

Make a 5x5 edge detection filter which is suited for the most cases (lowfreq_energy_ratio(array) > 0.97).

See also: FilterKernel, extract_edges.

CorrelationFunctions.Utilities.EdgeFilter — Type

EdgeFilter(bc, kernel)

Make an edge detection filter for extract_edges function. bc are boundary conditions of type BoundaryConditions and kernel is the kernel of type FilterKernel.

See also: extract_edges, BoundaryConditions, FilterKernel.